+++ src/main/java/org/folio/entlinks/domain/entity/AuthorityIdentifier.java
@@ -3,6 +3,7 @@ package org.folio.entlinks.domain.entity;
 import java.io.Serializable;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -13,6 +14,7 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
+@EqualsAndHashCode
 public class AuthorityIdentifier implements Serializable {
 
   private String value;